+++ src/BaseBinaryStar.h
@@ -391,6 +391,10 @@ private:
     double              m_ZetaLobe;
     double              m_ZetaStar;
 
+    // thresholds flags for system detailed output file
+    DBL_VECTOR          m_DetailedOutputAgeFlags1;
+    DBL_VECTOR          m_DetailedOutputAgeFlags2;
+    BOOL_VECTOR         m_DetailedOutputTimeFlags;
 
     // Binaries contain two stars
     BinaryConstituentStar *m_Star1;                                                         // Initially more massive star - the primary
@@ -529,8 +533,12 @@ private:
 
     bool PrintRLOFParameters(const RLOF_RECORD_TYPE p_RecordType = RLOF_RECORD_TYPE::DEFAULT);
     
-    bool PrintBinarySystemParameters(const BSE_SYSPARMS_RECORD_TYPE p_RecordType = BSE_SYSPARMS_RECORD_TYPE::DEFAULT) const { 
+    bool PrintSystemParameters(const BSE_SYSPARMS_RECORD_TYPE p_RecordType = BSE_SYSPARMS_RECORD_TYPE::DEFAULT) const { 
         return LOGGING->LogBSESystemParameters(this, p_RecordType);
+    }
+    
+    bool PrintSystemDetailedOutput(const BSE_SYSTEM_DETAILED_RECORD_TYPE p_RecordType = BSE_SYSTEM_DETAILED_RECORD_TYPE::DEFAULT) const { 
+        return LOGGING->LogBSESystemDetailedOutput(this, p_RecordType);
     }
     
     bool PrintDetailedOutput(const long int p_Id, const BSE_DETAILED_RECORD_TYPE p_RecordType) const {